Add unit test and clean others
