+++ cgmes/cgmes-gl/src/test/resources/importGL_GL.xml
@@ -1,0 +1,57 @@
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<rdf:RDF  xmlns:cim="http://iec.ch/TC57/2013/CIM-schema-cim16#" 
+      xmlns:md="http://iec.ch/TC57/61970-552/ModelDescription/1#" 
+      xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#">
+  <md:FullModel rdf:about="urn:uuid:fba05f4d-eda1-4c70-a042-d03acfba2eb3">
+    <md:Model.DependentOn rdf:resource="urn:uuid:94f88e95-1d5a-4227-984b-84232f4dfb81"/>
+  </md:FullModel>
+  <cim:CoordinateSystem rdf:ID="_wgs84">
+    <cim:CoordinateSystem.crsUrn>urn:ogc:def:crs:EPSG::4326</cim:CoordinateSystem.crsUrn>
+  </cim:CoordinateSystem>
+  <cim:Location rdf:ID="_locationLine">
+    <cim:Location.CoordinateSystem rdf:resource="#_wgs84" />
+    <cim:Location.PowerSystemResources rdf:resource="#_ACLine" />
+  </cim:Location>
+  <cim:Location rdf:ID="_location1">
+    <cim:Location.CoordinateSystem rdf:resource="#_wgs84" />
+    <cim:Location.PowerSystemResources rdf:resource="#_Substation1" />
+  </cim:Location>
+  <cim:Location rdf:ID="_location2">
+    <cim:Location.CoordinateSystem rdf:resource="#_wgs84" />
+    <cim:Location.PowerSystemResources rdf:resource="#_Substation2" />
+  </cim:Location>
+  <cim:PositionPoint rdf:ID="_positionLine0">
+    <cim:PositionPoint.Location rdf:resource="#_locationLine" />
+    <cim:PositionPoint.sequenceNumber>0</cim:PositionPoint.sequenceNumber>
+    <cim:PositionPoint.xPosition>-1.8468419</cim:PositionPoint.xPosition>
+    <cim:PositionPoint.yPosition>48.7123755</cim:PositionPoint.yPosition>
+  </cim:PositionPoint>
+  <cim:PositionPoint rdf:ID="_positionLine1">
+    <cim:PositionPoint.Location rdf:resource="#_locationLine" />
+    <cim:PositionPoint.sequenceNumber>1</cim:PositionPoint.sequenceNumber>
+    <cim:PositionPoint.xPosition>-1.8330389</cim:PositionPoint.xPosition>
+    <cim:PositionPoint.yPosition>48.8725425</cim:PositionPoint.yPosition>
+  </cim:PositionPoint>
+  <cim:PositionPoint rdf:ID="_positionLine2">
+    <cim:PositionPoint.Location rdf:resource="#_locationLine" />
+    <cim:PositionPoint.sequenceNumber>2</cim:PositionPoint.sequenceNumber>
+    <cim:PositionPoint.xPosition>-2.1263669</cim:PositionPoint.xPosition>
+    <cim:PositionPoint.yPosition>49.1703925</cim:PositionPoint.yPosition>
+  </cim:PositionPoint>
+  <cim:PositionPoint rdf:ID="_positionLine3">
+    <cim:PositionPoint.Location rdf:resource="#_locationLine" />
+    <cim:PositionPoint.sequenceNumber>3</cim:PositionPoint.sequenceNumber>
+    <cim:PositionPoint.xPosition>-2.136596</cim:PositionPoint.xPosition>
+    <cim:PositionPoint.yPosition>49.224448</cim:PositionPoint.yPosition>
+  </cim:PositionPoint>
+  <cim:PositionPoint rdf:ID="_positionSubstation1">
+    <cim:PositionPoint.Location rdf:resource="#_location1" />
+    <cim:PositionPoint.xPosition>-1.8468419</cim:PositionPoint.xPosition>
+    <cim:PositionPoint.yPosition>48.7123755</cim:PositionPoint.yPosition>
+  </cim:PositionPoint>
+  <cim:PositionPoint rdf:ID="_positionSubstation2">
+    <cim:PositionPoint.Location rdf:resource="#_location2" />
+    <cim:PositionPoint.xPosition>-2.136596</cim:PositionPoint.xPosition>
+    <cim:PositionPoint.yPosition>49.224448</cim:PositionPoint.yPosition>
+  </cim:PositionPoint>
+</rdf:RDF>
